+++ merqube_client_lib/types/__init__.py
@@ -1,8 +1,17 @@
 """
 Merqube custom types
 """
+from dataclasses import dataclass
 from enum import Enum, unique
 from typing import Any, Final
+
+import pandas as pd
+
+from merqube_client_lib.pydantic_types import (
+    EquityBasketPortfolio,
+    IdentifierUUIDPost,
+    IndexDefinitionPost,
+)
 
 HTTP_METHODS: Final[list[str]] = ["GET", "DELETE", "POST", "GET", "PUT", "OPTIONS", "PATCH"]
 
@@ -22,3 +31,12 @@ Manifest = dict[str, Any]
 ManifestList = list[Manifest]
 
 Records = list[dict[str, str]]
+
+TargetPortfoliosDates = list[tuple[pd.Timestamp, EquityBasketPortfolio]]
+
+
+@dataclass
+class CreateReturn:
+    template: IndexDefinitionPost
+    ident: IdentifierUUIDPost | None = None
+    initial_target_ports: TargetPortfoliosDates | None = None